+++ __tests/wow_api/TableUtil.lua
@@ -1,11 +1,26 @@
-local function ipairs_reverse_iter(table, index)
-	index = index - 1;
-	if index > 0 then
-		return index, table[index];
-	end
-end
 function ipairs_reverse(table)
-	return ipairs_reverse_iter, table, #table + 1;
+	local function Enumerator(table, index)
+		index = index - 1;
+		local value = table[index];
+		if value ~= nil then
+			return index, value;
+		end
+	end
+	return Enumerator, table, #table + 1;
+end
+function CreateTableEnumerator(tbl, indexBegin, indexEnd)
+	indexBegin = indexBegin and (indexBegin - 1) or 0;
+	indexEnd = indexEnd or math.huge;
+	local function Enumerator(tbl, index)
+		index = index + 1;
+		if index <= indexEnd then
+			local value = tbl[index];
+			if value ~= nil then
+				return index, value;
+			end
+		end
+	end
+	return Enumerator, tbl, indexBegin;
 end
 function tDeleteItem(tbl, item)
 	local size = #tbl;
@@ -26,7 +41,12 @@ function tIndexOf(tbl, item)
 	end
 end
 function tContains(tbl, item)
-	return tIndexOf(tbl, item) ~= nil;
+	for k, v in pairs(tbl) do
+		if item == v then
+			return true;
+		end
+	end
+	return false;
 end
 -- This is a deep compare on the values of the table (based on depth) but not a deep comparison
 -- of the keys, as this would be an expensive check and won't be necessary in most cases.
@@ -108,12 +128,17 @@ function MergeTable(destination, source)
 		destination[k] = v;
 	end
 end
-function AccumulateIf(tbl, pred)
+function Accumulate(tbl)
 	local count = 0;
 	for k, v in pairs(tbl) do
-		if pred(v) then
-			count = count + 1;
-		end
+		count = count + v;
+	end
+	return count;
+end
+function AccumulateOp(tbl, op)
+	local count = 0;
+	for k, v in pairs(tbl) do
+		count = count + op(v);
 	end
 	return count;
 end
@@ -133,6 +158,9 @@ function FindInTableIf(tbl, pred)
 	end
 	return nil;
 end
+function TableHasAnyEntries(tbl)
+	return next(tbl) ~= nil;
+end
 function CopyValuesAsKeys(tbl)
 	local output = {};
 	for k, v in ipairs(tbl) do
@@ -147,4 +175,39 @@ function SafePack(...)
 end
 function SafeUnpack(tbl)
 	return unpack(tbl, 1, tbl.n);
+end
+function GetOrCreateTableEntry(table, key, defaultValue)
+	local currentValue = table[key];
+	if currentValue == nil then
+		if defaultValue ~= nil then
+			currentValue = defaultValue;
+		else
+			currentValue = {};
+		end
+		table[key] = currentValue;
+	end
+	return currentValue;
+end
+function GetOrCreateTableEntryByCallback(table, key, callback)
+	local currentValue = table[key];
+	if currentValue == nil then
+		currentValue = callback();
+		table[key] = currentValue;
+	end
+	return currentValue;
+end
+function GetRandomArrayEntry(array)
+	return array[math.random(1, #array)];
+end
+function GetKeysArray(tbl)
+	local keysArray = {};
+	for key in pairs(tbl) do
+		table.insert(keysArray, key);
+	end
+	return keysArray;
+end
+function SwapTableEntries(lhsTable, rhsTable, key)
+	local lhsValue = lhsTable[key];
+	lhsTable[key] = rhsTable[key];
+	rhsTable[key] = lhsValue;
 end